% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\docType{data}
\name{extreme_points}
\alias{extreme_points}
\title{The extreme points of Japan}
\format{
A four length list consisting of \link[sf:st_point]{st_point}
}
\usage{
extreme_points
}
\description{
A list of the east, west, south and north ends of
Japan's territory (including remote islands).
The northernmost position is what the government claims.
}
\examples{
extreme_points

require("purrr")
extreme_points \%>\%
map(
  ~ sf::st_sfc(.x, crs = 4326)) \%>\%
reduce(c)
}
\seealso{
\url{https://www.gsi.go.jp/KOKUJYOHO/center.htm}
}
\keyword{datasets}
