% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-jgd2011_bbox.R
\docType{data}
\name{jgd2011_bbox}
\alias{jgd2011_bbox}
\title{JGD2011 / Japan Plane Rectangular CS}
\format{
A tibble with 19 rows 3 variables:
}
\usage{
jgd2011_bbox
}
\description{
Japanese Geodetic Datum 2011
}
\examples{
jgd2011_bbox
}
\keyword{datasets}
