% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-utils.R
\docType{data}
\name{jp47prefectural_offices}
\alias{jp47prefectural_offices}
\title{Japan Prefectural Goverment Offices}
\format{
A \link[sf:sf]{sf} contains 2 column and 47 rows.
}
\usage{
jp47prefectural_offices
}
\description{
Locations of 47 government offices in Japan's prefectures.
}
\details{
The original file was downloaded from
\url{https://www.gsi.go.jp/KOKUJYOHO/center.htm},
which parses the PDF data and organizes the coordinates of the prefectural hall.
}
\examples{
require("sf")
jp47prefectural_offices
}
\keyword{datasets}
