% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dohunbyo.R
\name{parse_lon_dohunbyo}
\alias{parse_lon_dohunbyo}
\alias{parse_lat_dohunbyo}
\title{Parse longitude and latitude values in DMS}
\usage{
parse_lon_dohunbyo(longitude)

parse_lat_dohunbyo(latitude)
}
\arguments{
\item{longitude}{longitude values}

\item{latitude}{latitude values}
}
\value{
\emph{numeric} vector
}
\description{
Parse longitude and latitude values in DMS
}
\examples{
x <- "\u6771\u7d4c139\u5ea644\u520628\u79d28869"
parse_lon_dohunbyo(x)
y <- "\u5317\u7def35\u5ea639\u520629\u79d21572"
parse_lat_dohunbyo(y)
}
