% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dohunbyo.R
\name{replace_dohunbyo_kanji}
\alias{replace_dohunbyo_kanji}
\title{Replace Kanji in degrees, minutes, and seconds with symbols}
\usage{
replace_dohunbyo_kanji(x)
}
\arguments{
\item{x}{character}
}
\value{
\emph{character} vector
}
\description{
Replace Kanji in degrees, minutes, and seconds with symbols
}
\examples{
x <- "\u6771\u7d4c139\u5ea644\u520628\u79d28869"
replace_dohunbyo_kanji(x)
y <- "\u5317\u7def35\u5ea639\u520629\u79d21572"
replace_dohunbyo_kanji(y)
}
